// papertrlui-v2/src/app/shared/dto/vendor/vendor-bill-table-dto.ts
export class VendorBillTableDto {
  attachmentId: any;
  id: any;
  billNo: any;
  vendorId: any;
  billDate: any;
  billAmount: number;
  terms: string;
  dueDate: any;
  vendor: string;
  poNo: any;
  poReceipt: any;
  submittedBy: string;
  submittedDate: any;
  remainingBalance: any;
  discountAmount: any;
  approvalUser: string;
  approvalDate: any;
  status: string;
}